Struct recs::ComponentFilter

pub struct ComponentFilter {
    // some fields omitted
}

List of component types.

The Ecs methods has_all and collect_with each take a ComponentFilter instance. The recommended way to actually create a ComponentFilter is with the component_filter! macro.

Methods

impl ComponentFilter

fn new() -> Self

Create a new component filter.

fn add<C: Component>(&mut self)

Add component type C to the filter.

fn remove<C: Component>(&mut self)

Remove component type C from the filter.

fn contains<C: Component>(&mut self) -> bool

Return true if the filter already contains component type C; otherwise false.

fn from_slice(slice: &[TypeId]) -> Self

Create a component filter from a vector/slice of TypeId instances. (Not recommended; used by the component_filter! macro.)

fn iter<'a>(&'a self) -> Box<Iterator<Item=TypeId> + 'a>

Return an iterator over all the contained component types.