Crate recs

Modules

errors

Statics

DEBUGGING

Debugging should be set while initializing the lib, If no defined the default is disabled

PROGNAME

This value is set by set_prog it is used for logging creating paths and other functions. to handel its creation or modification use set_prog() to avoid wrapping

Functions

_get_array_props

check_map

decrypt_raw

encrypt_raw

initialize

Initialize checks the progname, and debugging values snf ensure the lib is ready to function

insert

Insert takes a relative path encrypts and stores files. Weather or not they’re deleted is based on values in the config.rs file

ping

remove

retrive

Retrieve starts a request to decrypt the file requested on sucess it returns where the file currently is ‘String’ and where to file was when it was encrypted ‘String’ it is up to the client to decide weather to move the file there or read the contents and delete the file

set_debug

Changes some mandatory logging functions and enables longer outputs in logs

set_prog

This function handels setting the PROGNAME variables

update_map